pyproject.toml
[project]
name = "sar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "torch>=2.8.0",
    "transformers>=4.56.2",
    "typer>=0.19.2",
    "wandb>=0.22.0",
]
